modularize client code to store containers and components separate
