+++ client/src/views/Components/index.js
@@ -1,0 +1,5 @@
+import { default as App } from './App';
+import { default as ErrorDisplay } from './ErrorDisplay';
+import { default as RadioButton } from './RadioButton';
+
+export { App, ErrorDisplay, RadioButton };
