+++ client/src/views/Containers/CurrentWeather.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
-import { weatherAppAPI } from '../utils';
+import { weatherAppAPI } from '../../utils';
 
 // Assets
-import ThunderStormIcon from './assets/weather_icons/01W.svg';
-import RainIcon from './assets/weather_icons/02W.svg';
-import SnowIcon from './assets/weather_icons/03W.svg';
-import ClearIcon from './assets/weather_icons/04W-DAY.svg';
-import CloudsIcon from './assets/weather_icons/05W.svg';
-import NoLocationFound from './assets/no-location.svg';
-import LoadingIcon from './assets/loading.svg';
+import ThunderStormIcon from '../assets/weather_icons/01W.svg';
+import RainIcon from '../assets/weather_icons/02W.svg';
+import SnowIcon from '../assets/weather_icons/03W.svg';
+import ClearIcon from '../assets/weather_icons/04W-DAY.svg';
+import CloudsIcon from '../assets/weather_icons/05W.svg';
+import NoLocationFound from '../assets/no-location.svg';
+import LoadingIcon from '../assets/loading.svg';
 
 class CurrentWeather extends Component {
   constructor(props) {
