+++ client/src/views/Containers/index.js
@@ -1,0 +1,4 @@
+import { default as Home } from './Home';
+import { default as CurrentWeather } from './CurrentWeather';
+
+export { Home, CurrentWeather };
